Refactor Effelsberg pipeline into modular package
